// client/src/features/SubAssets/types/asset.types.ts
export interface SubAsset {
  internalId: number;
  parentAssetId: string;
  name: string;
  assetId: string;
  btoAlignment: string | null;
  version: string | null;
  cmdbStatus: string;
  deploymentLifecyclePhase: string | null;
  applicationTypeFinancial: string | null;
  itOwner: string | null;
  businessOwner: string | null;
  businessOwnerSme: string | null;
  supportedBy: string | null;
  supportSme: string | null;
  architect: string | null;
  division: string | null;
  blockFundingName: string | null;
  blockFundingOwner: string | null;
  assessmentCategory: string | null;
  deploymentLifecycleStartDate: string | null;
  assetType: string | null;
  hosted: string | null;
  sox: string | null;
  customerFacing: string | null;
  sppi: string | null;
  ppiClassification: string | null;
  foundational: string | null;
  missionCritical: string | null;
  businessCritical: string | null;
  supporting: string | null;
  cotsOrInHouseBuilt: string | null;
  isSaas: string | null;
  maintenanceWindow: string | null;
  operationalHours: string | null;
  description: string | null;
  externalFacing: string | null;
  financialImpact4hrOutage: string | null;
  assetTier: string | null;
  informationClassification: string | null;
  createdAt: string;
  lastModifiedBy: string;
  lastModifiedDate: string;
  [key: string]: any;
}